sb pr float

+++ fb_app/models.py
@@ -653,11 +653,11 @@ class PlayoffPicks(models.Model):
     home_runner = models.PositiveIntegerField()
     home_receiver = models.PositiveIntegerField()
     home_passing = models.PositiveIntegerField(null=True)
-    home_passer_rating = models.PositiveIntegerField(default=100.0)
+    home_passer_rating = models.FloatField(default=100.0)
     away_runner = models.PositiveIntegerField()
     away_receiver = models.PositiveIntegerField()
     away_passing = models.PositiveIntegerField(null=True)
-    away_passer_rating = models.PositiveIntegerField(default=100.0)
+    away_passer_rating = models.FloatField(default=100.0)
     winning_team = models.ForeignKey(Teams, on_delete=models.CASCADE)
 
     def __str__(self):
